refactor: update topkio-gemini
