+++ topkio-gemini/src/gemini/data.rs
@@ -1,0 +1,51 @@
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize)]
+pub struct GeminiRequest {
+    pub contents: Vec<Content>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub safety_settings: Option<Vec<SafetySetting>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub generation_config: Option<GenerationConfig>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Content {
+    pub parts: Vec<Part>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub role: Option<String>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Part {
+    pub text: String,
+}
+
+#[derive(Debug, Serialize)]
+pub struct SafetySetting {
+    pub category: String,
+    pub threshold: String,
+}
+
+#[derive(Debug, Serialize)]
+pub struct GenerationConfig {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub temperature: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub top_p: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub max_output_tokens: Option<i32>,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct GeminiResponse {
+    pub candidates: Vec<Candidate>,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct Candidate {
+    pub content: Content,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub finish_reason: Option<String>,
+}
